cleanup(core): split MRTR types out of task_v2.go into mrtr.go
`core/task_v2.go` held both the SEP-2663 tasks-v2 wire types
(`TaskInfoV2`, `CreateTaskResult`, `DetailedTask`, etc.) and the
SEP-2322 MRTR types (`InputRequiredResult`, `requestState` signing
helpers, `MRTRRoundState`). A reader who lands on `task_v2.go`
looking for `InputRequiredResult` (correctly an SEP-2322 type) finds
the wrong file name.

Split by SEP boundary:

- `core/mrtr.go` (NEW, 327 lines) — `ResultType` discriminator,
  `InputRequest` / `InputRequests` / `InputResponses`,
  `InputRequiredResult`, the requestState error sentinels,
  `requestStatePayload` + `Sign/VerifyRequestState`, `MRTRRoundState`
  + `Sign/VerifyMRTRState`, `Encode/DecodeMRTRStatePlaintext`.
- `core/task_v2.go` (kept, 196 lines) — `TasksExtensionID` +
  `ClientSupportsTasks`, `TaskInfoV2`, `CreateTaskResult`,
  `DetailedTask` + narrowed aliases + `GetTaskResult`, `TaskError`,
  `UpdateTaskRequest` / `UpdateTaskResult`, `CancelTaskResult`.

The `ResultType` discriminator moves to `mrtr.go` because SEP-2322
owns it — tasks-v2 just consumes the same enum. Same logic for
`InputRequest` / `InputRequests` / `InputResponses`.

Tests split the same way:

- `core/mrtr_test.go` — `TestResultTypeConstants`, the four
  `TestInputRequest*` / `TestInputResponses*` cases, the two
  `TestInputRequiredResult*` cases, and the seven
  `TestRequestState_*` HMAC round-trip tests.
- `core/task_v2_test.go` — the SEP-2663 wire-shape tests for
  `TaskInfoV2`, `CreateTaskResult`, `DetailedTask` (working +
  completed + failed + input_required variants), `UpdateTaskRequest`,
  and `TestAckShapes`.

Pure mechanical move. No behavior change. No public-API rename.
Same `core` package, so no external import statements change. The
git rename mechanic was `git mv core/task_v2.go core/mrtr.go`
(plus the test analog), preserving `git log --follow` traceback for
the MRTR side (which got the majority of the original content). The
new `core/task_v2.go` is a fresh file from this commit.

Addresses issue 433 item: "Split `core/task_v2.go` into `core/task_v2.go`
+ `core/mrtr.go`". The other three items on 433 stay open.
